PizzaHat/requirements.txt
git+https://github.com/Rapptz/discord.py
asyncpg
TagScriptEngine
aiohttp
asyncio
pyparsing
python-dotenv
typing
typing-extensions
asyncdagpi
requests
humanfriendly
discord-together
numpy
jishaku
psutil
topggpy